Move depSolver checks back to depCheck.go
